requirements.txt
absl-py==0.11.0
astunparse==1.6.3
autopep8==1.5.4
black==19.10b0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
contextvars==2.4
Flask==1.1.2
Flask-Cors==3.0.9
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googletrans==3.0.0
grpcio==1.33.2
gTTS==2.2.1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.14
importlib-metadata==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
mpyg321==0.1.0
numpy==1.19.3
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opt-einsum==3.3.0
pathspec==0.8.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pygame==2.0.0
pymongo==3.11.1
pypiwin32==223
pytesseract==0.3.6
pyttsx3==2.90
pywin32==300
regex==2020.6.8
requests==2.25.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sniffio==1.2.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
